MainMenu, PreGame, NPC DialogueSOs
+ Main Menu Functions done
Desc: I've finished the main menu for the most part but it still needs reformatting (placements and sizes need to be altered).
+ Pre Game Scene done
Desc: I've finished the pregame scene where the we ask for the player's username and save it
+ NPCDialogueSOs
Desc: I've mostly thought about how I'm gonna structure the dialogues and whatnot and created a Scriptable Object based off of it that we can use for every dialogue later

Additional Notes:
I need to somehow make all the {PlayerName}-s in the SO be whatever it is in the playerprefs.